Add minimal git config to repos
The test don't pass unless a minimal configuration for Git is provided.
With this change, we don't rely on the final user to provide this
configuration.

Every time a repository is cloned on tests we're adding a configuration
only for that repository that's not going to clash with any other
configuration already set.